% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc_charts.R
\name{add_citation}
\alias{add_citation}
\title{Helper - add source citation to ggplot object}
\usage{
add_citation(chart, source)
}
\arguments{
\item{chart}{ggplot or plotly object}

\item{source}{Source reference as character string}
}
\description{
Helper - add source citation to ggplot object
}
