% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc_line_charts.R
\name{cleveland_dot_chart}
\alias{cleveland_dot_chart}
\title{Cleveland Dot Chart}
\usage{
cleveland_dot_chart(
  t,
  x,
  y,
  fill,
  est = NULL,
  dec = 0,
  dform = "\%b-\%Y",
  lwidth = 1.5,
  dotsize = 3,
  color = "gnbopgy_5",
  shape = "circle",
  title = NULL,
  subtitle = NULL,
  source = "",
  alt = NULL,
  xlabel = NULL,
  ylabel = NULL
)
}
\arguments{
\item{t}{A tibble or dataframe in long form for plotting}

\item{x}{The name of the variable you want plotted on the X-Axis (categorical)}

\item{y}{The name of the variable you want plotted on the Y-Axis (numeric)}

\item{fill}{The name of the variable supplying line endpoints (e.g. time points)}

\item{est}{Type for the numeric values - enter "percent", "currency" or "number", defaults to "percent"}

\item{dec}{Number of decimal points in labels - defaults to 0}

\item{lwidth}{Width of lines; defaults to 1.5}

\item{dotsize}{Defaults to 3}

\item{color}{Name of color palette to use - generally defaults to "pgnobgy_5"}

\item{shape}{Of dot and legend; can be either value 1:25 or name; defaults to "circle"
see https://ggplot2.tidyverse.org/articles/ggplot2-specs.html?q=shapes#sec:shape-spec}

\item{title}{Title to be used for chart, if desired - defaults to "NULL"}

\item{subtitle}{Sub-title to be used for chart, if desired - defaults to "NULL"}

\item{source}{Source reference as character string, if desired - defaults to blank}
}
\value{
A Cleveland line chart
}
\description{
Creates lines showing range by category, with dot endpoints
credit https://uc-r.github.io/cleveland-dot-plots
}
