% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc_charts.R
\name{column_bar_charts}
\alias{column_bar_charts}
\alias{static_column_chart}
\alias{static_bar_chart}
\alias{interactive_column_chart}
\alias{interactive_bar_chart}
\title{Column and bar charts}
\usage{
static_column_chart(t, x, y, fill, xlabel = NULL, ylabel = NULL, ...)

static_bar_chart(t, x, y, fill, xlabel = NULL, ylabel = NULL, ...)

interactive_column_chart(t, x, y, fill, xlabel = NULL, ylabel = NULL, ...)

interactive_bar_chart(t, x, y, fill, xlabel = NULL, ylabel = NULL, ...)
}
\arguments{
\item{t}{A tibble in long form for plotting}

\item{x}{The name of the variable you want plotted on the X-Axis}

\item{y}{The name of the variable you want plotted on the Y-Axis}

\item{fill}{The name of the variable you want the fill color of the bars to be based on}

\item{xlabel}{category-axis title to be used for chart, if desired - defaults to "NULL"}

\item{ylabel}{numeric-axis title to be used for chart, if desired - defaults to "NULL"}

\item{...}{additional arguments passed to  \code{\link{generic_column_bar}}}
}
\value{
static or interactive column or bar chart
}
\description{
Separate functions for static and interactive, column and bar
}
