% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc_charts.R
\name{create.line.chart}
\alias{create.line.chart}
\title{Create PSRC Line Chart}
\usage{
create.line.chart(
  t,
  w.x,
  w.y,
  w.g,
  w.title = NULL,
  w.sub.title = NULL,
  est.type = "percent",
  w.dec = 0,
  x.type = "Date",
  d.form = "\%b-\%Y",
  w.breaks = NULL,
  w.lwidth = 1,
  w.color = "psrc_light"
)
}
\arguments{
\item{t}{A tibble or dataframe in long form for plotting}

\item{w.x}{The name of the variable you want plotted on the X-Axis}

\item{w.y}{The name of the variable you want plotted on the Y-Axis}

\item{w.g}{The name of the variable you want to fill the bars}

\item{w.title}{Title to be used for chart, if desired - defaults to "NULL"}

\item{w.sub.title}{Sub-title to be used for chart, if desired - defaults to "NULL"}

\item{est.type}{Type for the Y values - enter "percent", "currency" or "number", defaults to "percent"}

\item{w.dec}{Number of decimal points in labels - defaults to 0}

\item{x.type}{Type of Data being plotted on the X-Axis - enter "Date" or "Continuous", defaults "Date"}

\item{d.form}{If the X-axis is for Dates, format of labels for Axis, defaults to "%b-%Y"}

\item{w.breaks}{If the X-axis is Continuous, define the number of breaks- defaults to "NULL"}

\item{w.lwidth}{Line width - defaults to "1"}

\item{w.color}{Name of color palette to use - defaults to "psrc_light"}
}
\value{
static line chart
}
\description{
This function allows you to create line charts.
}
\examples{
\dontrun{
library(dplyr)

# Read in the example data and filter to Mode to Work for Everyone for all years in the data
df <- read.csv("inst/extdata/example_data.csv") \%>\% 
      filter(Category=="Mode to Work by Race") \%>\%
      filter(Geography=="Region" & Race=="Total") \%>\%
      mutate(Year = as.character(Year)) \%>\%
      filter(Year=="2020")

my.chart <- create_line_chart(t=seatac.airport, w.x='day', w.y='passenger_screenings', w.g='year', est.type="number", 
                              w.title="Weekly Passsenger Screenings: 2019 to 2022", d.form = "\%B", w.lwidth=3)

}

}
