% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc_charts.R
\name{create_bar_chart}
\alias{create_bar_chart}
\title{Create Bar Chart}
\usage{
create_bar_chart(
  t,
  x,
  y,
  f,
  href = NULL,
  hrefnm = NULL,
  hrefcl = NULL,
  moe = NULL,
  pos = "dodge",
  est = "percent",
  title = "",
  subtitle = "",
  source = "",
  dec = 0,
  color = "psrc_dark",
  interactive = "no"
)
}
\arguments{
\item{t}{A tibble in long form for plotting}

\item{x}{The name of the variable you want plotted on the X-Axis}

\item{y}{The name of the variable you want plotted on the Y-Axis}

\item{f}{The name of the variable you want the fill color of the bars to be based on}

\item{href}{A list of values to be used for any horizontal reference lines - default is "NULL"}

\item{hrefnm}{A list of names to be used for any horizontal reference lines that is equal length to the number of lines - default is "NULL"}

\item{hrefcl}{A list of colors to be used for any horizontal reference lines that is equal length to the number of lines - default is "NULL"}

\item{moe}{The name of the variable to be used for error bars, if desired - default is "NULL"}

\item{pos}{Determines if the bars are side-by-side(dodge) or stacked(stack) - defaults to "dodge"}

\item{est}{Type for the Y values - enter "percent", "currency" or "number", defaults to "percent"}

\item{title}{Title to be used for chart, if desired - defaults to "NULL"}

\item{subtitle}{Sub-title to be used for chart, if desired - defaults to "NULL"}

\item{source}{Source reference to be used for chart, if desired - defaults to "NULL"}

\item{dec}{Number of decimal points in labels - defaults to 0}

\item{color}{Name of color palette to use - defaults to "psrc_dark"}

\item{interactive}{Enable hover text and other interactive features - defaults to "no"}
}
\value{
bar (horizontal bar) chart that is either static or interactive depending on choice
}
\description{
This function allows you to create bar charts (horizontal bars).
}
\examples{

library(psrcplot)
library(psrctrends)
library(dplyr)

install_psrc_fonts()

ev_registration <- get_ev_registration_ytd()

tbl <- ev_registration \%>\% 
       filter(county=="PSRC Region")

static_chart <- create_bar_chart(t=tbl, 
                                 x="date", y="share", f="variable",
                                 title="Year to Date Share of Vehicle Registrations", 
                                 source="Source: Washington State Open Data Portal",
                                 pos = "stack")
                                   
interactive_chart <- create_bar_chart(t=tbl, 
                                      x="date", y="share", f="variable",
                                      title="Year to Date Share of Vehicle Registrations", 
                                      source="Source: Washington State Open Data Portal",
                                      pos = "stack", interactive='yes')                                    
                                  

}
