% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc_charts.R
\name{create_bubble_chart}
\alias{create_bubble_chart}
\title{Create PSRC Bubble Chart}
\usage{
create_bubble_chart(
  t,
  w.x,
  w.y,
  f,
  s,
  w.color = "psrc_light",
  w.title = NULL,
  w.sub.title = NULL
)
}
\arguments{
\item{t}{A tibble or dataframe in long form for plotting}

\item{w.x}{The name of the variable you want plotted on the X-Axis}

\item{w.y}{The name of the variable you want plotted on the Y-Axis}

\item{f}{The name of the variable you want the fill color of the bubbles to be based on}

\item{s}{The name of the variable used to size the bubbles}

\item{w.color}{Name of color palette to use - defaults to "psrc_light"}

\item{w.title}{Title to be used for chart, if desired - defaults to "NULL"}

\item{w.sub.title}{Sub-title to be used for chart, if desired - defaults to "NULL"}
}
\value{
bubble chart
}
\description{
This function allows you to create a bubble charts.
}
\examples{

library(dplyr)
library(psrctrends)
library(psrcplot)

# Pull NTD Transit Data by Urbanized Area
uza.data <- process_ntd_uza_data(yr="2022", census.yr="2020")

# Create Bubble chart for Transit Boardings per Capita
uza.chart <- create_bubble_chart(t=uza.data \%>\% dplyr::filter(variable == "Boardings per Capita"),
                                 w.x="population", w.y="estimate", f <- "plot_id", 
                                 s <- "population", w.color <- "psrc_light",
                                 w.title = "Boardings per Capita", 
                                 w.sub.title = "Urbanized Areas with at least 1 million people")

}
