% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc_charts.R
\name{create_column_chart}
\alias{create_column_chart}
\title{Create Column Chart}
\usage{
create_column_chart(
  t,
  x,
  y,
  f,
  pos = "dodge",
  est = "percent",
  moe = NULL,
  href = NULL,
  hrefnm = NULL,
  hrefcl = NULL,
  title = NULL,
  subtitle = NULL,
  source = NULL,
  dec = 0,
  color = "psrc_dark",
  interactive = "no"
)
}
\arguments{
\item{t}{A tibble in long form for plotting}

\item{x}{The name of the variable you want plotted on the X-Axis}

\item{y}{The name of the variable you want plotted on the Y-Axis}

\item{f}{The name of the variable you want the fill color of the bars to be based on}

\item{pos}{Determines if the bars are side-by-side(dodge) or stacked(stack) - defaults to "dodge"}

\item{est}{Type for the Y values - enter "percent", "currency" or "number", defaults to "percent"}

\item{moe}{The name of the variable to be used for error bars, if desired - default is "NULL"}

\item{href}{A list of values to be used for any horizontal reference lines - default is "NULL"}

\item{hrefnm}{A list of names to be used for any horizontal reference lines that is equal length to the number of lines - default is "NULL"}

\item{hrefcl}{A list of colors to be used for any horizontal reference lines that is equal length to the number of lines - default is "NULL"}

\item{title}{Title to be used for chart, if desired - defaults to "NULL"}

\item{subtitle}{Sub-title to be used for chart, if desired - defaults to "NULL"}

\item{source}{Source reference to be used for chart, if desired - defaults to "NULL"}

\item{dec}{Number of decimal points in labels - defaults to 0}

\item{color}{Name of color palette to use - defaults to "psrc_dark"}

\item{interactive}{Enable hover text and other interactive features - defaults to "no"}
}
\value{
column (vertical bar) chart that is either static or interactive depending on choice
}
\description{
This function allows you to create column charts (vertical bars).
}
\examples{

library(dplyr)
modes_shares <- read.csv(system.file('extdata', 'example_data.csv', package='psrcplot')) \%>\% 
 filter(Category=="Mode to Work by Race") \%>\%
 filter(Geography=="Region" & Race=="Total") \%>\%
 mutate(Year = as.character(Year))

# Static Chart with Error Bars         
static_chart <- create_column_chart(t=modes_shares, x="Mode",y="share", f="Year",
                                    title="Mode Share to Work",
                                    source="Source: ACS 5yr Data Table B3002",
                                    moe="share_moe", color="psrc_light")

# Interactive Chart without Error Bars         
interactive_chart <- create_column_chart(t=modes_shares, x="Mode",y="share", f="Year",
                                         title="Mode Share to Work",
                                         source="Source: ACS 5yr Data Table B3002",
                                         color="psrc_light", interactive='yes')
                                         
}
