% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc_charts.R
\name{create_facet_bar_chart}
\alias{create_facet_bar_chart}
\title{Create PSRC Facet Bar Charts}
\usage{
create_facet_bar_chart(
  t,
  w.x,
  w.y,
  f,
  g,
  w.moe = NULL,
  est.type = "percent",
  w.scales = "free",
  w.facet = 3,
  w.dec = 0,
  l.pos = "above",
  w.color = "psrc_dark",
  w.title = NULL,
  w.sub.title = NULL,
  w.interactive = "no"
)
}
\arguments{
\item{t}{A tibble or dataframe in long form for plotting}

\item{w.x}{The name of the variable you want plotted on the X-Axis}

\item{w.y}{The name of the variable you want plotted on the Y-Axis}

\item{f}{The name of the variable you want the fill color of the bars to be based on}

\item{g}{The name of the variable to be the facets}

\item{w.moe}{The name of the variable to be used for error bars, if desired - default to "NULL"}

\item{est.type}{Type for the Y values - enter "percent", "currency" or "number", defaults to "percent"}

\item{w.scales}{Value for axis in facets, either "fixed" or "free" - defaults to "free"}

\item{w.facet}{Value for the number of columns in your facet - defaults to 3}

\item{w.dec}{Number of decimal points in labels - defaults to 0}

\item{l.pos}{Position for the bar labels of "above" or "within" - defaults to "above"}

\item{w.color}{Name of color palette to use - defaults to "psrc_dark"}

\item{w.title}{Title to be used for chart, if desired - defaults to "NULL"}

\item{w.sub.title}{Sub-title to be used for chart, if desired - defaults to "NULL"}

\item{w.interactive}{Enable hover text and other interactive features - defaults to "no"}
}
\value{
facet bar chart that is either static or interactive depending on choice
}
\description{
This function allows you to create facet bar charts.
}
\examples{

library(dplyr)
# Read in the example data and filter to 2020 Population by Race by County
df <- read.csv(system.file('extdata', 'example_data.csv', package='psrcplot')) \%>\% 
      filter(Category=="Population by Race" & Year==2020) \%>\%
      filter(Geography!="Region" & Race!="Total")

# Create a facet chart for population by race using counties as the facet
my.chart <- create_facet_bar_chart(t=df, w.x="Race", w.y="share", 
                                   f="Race", g="Geography", 
                                   w.facet=2, w.scales="fixed")

}
