% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc_charts.R
\name{create_treemap_chart}
\alias{create_treemap_chart}
\title{Create PSRC TreeMap Chart}
\usage{
create_treemap_chart(
  t,
  s,
  fill,
  title = NULL,
  subtitle = NULL,
  est = "percent",
  dec = 0,
  color = "psrc_light"
)
}
\arguments{
\item{t}{A tibble or dataframe in long form for plotting}

\item{s}{The name of the variable with the value you want to use to size the bars}

\item{fill}{The name of the variable you want to fill the bars}

\item{title}{Title to be used for chart, if desired - defaults to "NULL"}

\item{subtitle}{Sub-title to be used for chart, if desired - defaults to "NULL"}

\item{est}{Type for the Y values - enter "percent", "currency" or "number", defaults to "percent"}

\item{dec}{Number of decimal points in labels - defaults to 0}

\item{color}{Name of color palette to use - defaults to "psrc_light"}
}
\value{
static treemap chart
}
\description{
This function allows you to create treemap charts.
}
\examples{

library(dplyr)

# Read in the example data and filter to Mode to Work for Everyone for all years in the data
df <- read.csv(system.file('extdata', 'example_data.csv', package='psrcplot')) \%>\% 
      filter(Category=="Mode to Work by Race") \%>\%
      filter(Geography=="Region" & Race=="Total") \%>\%
      mutate(Year = as.character(Year)) \%>\%
      filter(Year=="2020")

my.chart <- create_treemap_chart(t=df, s="share", fill="Mode", title="Mode Share to Work")


}
