% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc_echarts.R
\name{echart_bar_chart}
\alias{echart_bar_chart}
\title{Create an echart4r bar or column chart}
\usage{
echart_bar_chart(
  df,
  category_var,
  numeric_var,
  fill,
  pos = "NULL",
  column_vs_bar = "column",
  est = "percent",
  title
)
}
\arguments{
\item{df}{}

\item{category_var}{}

\item{numeric_var}{}

\item{fill}{}

\item{pos}{The position either NULL or "grp" (stacked)}

\item{column_vs_bar}{"column" or "bar"}

\item{est}{Estimate type, enter "percent", "currency" or "number", defaults to "percent"}

\item{title}{}
}
\value{
An interactive bar or column chart via echarts4r
}
\description{
Create an echart4r bar or column chart
}
\examples{
library(tidyverse)

d <- mode_share_example_data \%>\% 
filter(Year == 2020)

echart_bar_chart(df = d ,
                category_var = "Race",
                numeric_var = "share",
                fill = "Geography",
                title = "Test Chart",
                pos = NULL,
                column_vs_bar = "bar",
                est = "number"
)
}
