% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc_echarts.R
\name{echart_line_chart}
\alias{echart_line_chart}
\title{echart4r Line Chart}
\usage{
echart_line_chart(t, x, y, ...)
}
\arguments{
\item{t}{A data frame in long form for plotting}

\item{x}{The name of the category variable}

\item{y}{The name of the variable with numeric values to plot}

\item{...}{Additional arguments passed to \code{\link{generic_echart}}}
}
\value{
An interactive line chart via echarts4r
}
\description{
echart4r Line Chart
}
