% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrcplot_example_data.R
\docType{data}
\name{enplanement_example_data}
\alias{enplanement_example_data}
\title{Airport example data included in psrcplot}
\format{
An object of class \code{data.frame} with 26 rows and 4 columns.
}
\usage{
enplanement_example_data
}
\description{
SELECT ad.loc_id, ad.airport_name, ef.data_year, ef.enplanements
FROM Elmer.faa.airport_dims AS ad JOIN Elmer.faa.enplanement_facts AS ef 
ON ad.airport_dim_id=ef.airport_dim_id
WHERE ad.st='WA' AND ad.airport_dim_id IN(657,671)
ORDER BY ad.airport_name, ef.data_year;
}
\keyword{datasets}
