% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrcplot_helpers.R
\name{est_label_formats}
\alias{est_label_formats}
\title{Helper - return label format function to match estimate type}
\usage{
est_label_formats(est)
}
\arguments{
\item{est}{Type for the numeric values - enter "percent", "currency" or "number", defaults to "percent"}
}
\description{
Helper - return label format function to match estimate type
}
