% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrcplot_helpers.R
\name{est_number_formats}
\alias{est_number_formats}
\title{Helper - return number formatting to match estimate type}
\usage{
est_number_formats(est)
}
\arguments{
\item{est}{Type for the numeric values - enter "percent", "currency" or "number"}
}
\description{
Helper - return number formatting to match estimate type
}
