% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrcplot_helpers.R
\name{est_type_default}
\alias{est_type_default}
\title{Helper - return estimate type if null}
\usage{
est_type_default(x)
}
\arguments{
\item{x}{numeric vector or dataframe column}
}
\description{
Helper - return estimate type if null
}
