% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc_charts.R
\name{generic_column_bar}
\alias{generic_column_bar}
\title{Generic column/bar workhorse function}
\usage{
generic_column_bar(
  t,
  category_var,
  numeric_var,
  fill,
  pos = "dodge",
  est = "percent",
  moe = NULL,
  href = NULL,
  hrefnm = NULL,
  hrefcl = NULL,
  title = NULL,
  subtitle = NULL,
  source = "",
  alt = NULL,
  category_label = NULL,
  numeric_label = NULL,
  axis_scale = 1,
  column_vs_bar = "column",
  dec = 0,
  color = "pgnobgy_5",
  interactive = FALSE
)
}
\arguments{
\item{t}{A tibble in long form for plotting}

\item{category_var}{The name of the category variable}

\item{numeric_var}{The name of the variable with numeric values to plot}

\item{fill}{The name of the variable you want the fill color of the bars to be based on}

\item{pos}{Determines if the bars are side-by-side(dodge) or stacked(stack) - defaults to "dodge"}

\item{est}{Type for the numeric values - enter "percent", "currency" or "number", defaults to "percent"}

\item{moe}{The name of the variable to be used for error bars, if desired - default is "NULL"}

\item{href}{A list of values to be used for any horizontal reference lines - default is "NULL"}

\item{hrefnm}{A list of names to be used for any horizontal reference lines that is equal length to the number of lines - default is "NULL"}

\item{hrefcl}{A list of colors to be used for any horizontal reference lines that is equal length to the number of lines - default is "NULL"}

\item{title}{Title to be used for chart, if desired - defaults to "NULL"}

\item{subtitle}{Sub-title to be used for chart, if desired - defaults to "NULL"}

\item{source}{Source reference as character string, if desired - defaults to blank}

\item{alt}{Text to be used for alt-text, if desired - defaults to "NULL"}

\item{category_label}{category-axis title to be used for chart, if desired - defaults to "NULL"}

\item{numeric_label}{numeric-axis title to be used for chart, if desired - defaults to "NULL"}

\item{axis_scale}{Enlarge or reduce the axis weight - defaults to 1}

\item{column_vs_bar}{"column": vertical bars or "bar": horizontal bars - defaults to "column"}

\item{dec}{Number of decimal points in labels - defaults to 0}

\item{color}{Name of color palette to use - generally defaults to "pgnobgy_5"}

\item{interactive}{Enable hover text and other interactive features - defaults to FALSE}
}
\value{
static or interactive column or bar chart
}
\description{
Helper function to \code{\link{column_bar_charts}}
}
