% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc_fonts.R
\name{install_psrc_fonts}
\alias{install_psrc_fonts}
\title{Install PSRC Fonts
Installs Google Fonts used in PSRC Visualizations}
\usage{
install_psrc_fonts()
}
\description{
This function will install fonts from Google for use in PSRC visualizations.
}
\examples{

install_psrc_fonts()

}
