% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc_charts.R
\name{interactive_bar_chart}
\alias{interactive_bar_chart}
\title{Create an Interactive Bar Chart}
\usage{
interactive_bar_chart(
  t,
  x,
  y,
  fill,
  pos = "dodge",
  est = "percent",
  moe = NULL,
  href = NULL,
  hrefnm = NULL,
  hrefcl = NULL,
  title = NULL,
  dec = 0,
  color = "pgnobgy_5"
)
}
\arguments{
\item{t}{A tibble in long form for plotting}

\item{x}{The name of the variable you want plotted on the X-Axis}

\item{y}{The name of the variable you want plotted on the Y-Axis}

\item{fill}{The name of the variable you want the fill color of the bars to be based on}

\item{pos}{Determines if the bars are side-by-side(dodge) or stacked(stack) - defaults to "dodge"}

\item{est}{Type for the Y values - enter "percent", "currency" or "number", defaults to "percent"}

\item{moe}{The name of the variable to be used for error bars, if desired - default is "NULL"}

\item{href}{A list of values to be used for any horizontal reference lines - default is "NULL"}

\item{hrefnm}{A list of names to be used for any horizontal reference lines that is equal length to the number of lines - default is "NULL"}

\item{hrefcl}{A list of colors to be used for any horizontal reference lines that is equal length to the number of lines - default is "NULL"}

\item{title}{Title to be used for chart, if desired - defaults to "NULL"}

\item{dec}{Number of decimal points in labels - defaults to 0}

\item{color}{Name of color palette to use - defaults to "pgnobgy_5"}
}
\value{
interactive bar (horizontal bar) chart
}
\description{
This function allows you to create bar charts (horizontal bars).
}
\examples{

library(dplyr)
library(stringr)
mode_shares <- read.csv(system.file('extdata', 'example_data.csv', package='psrcplot')) \%>\% 
  filter(Category=="Mode to Work by Race") \%>\%
  filter(Geography=="Region" & Race=="Total") \%>\%
  mutate(Year = as.character(Year)) \%>\%
  mutate(axis_labels = str_wrap(Mode, width = 10))

modes_bar_chart <- interactive_bar_chart(t=mode_shares, x="share", y="axis_labels", fill="Year",
                                         est = "percent",
                                         title="Mode Share to Work",
                                         color="gnbopgy_5", dec = 0)
}
