% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc_charts.R
\name{interactive_line_chart}
\alias{interactive_line_chart}
\title{Interactive Line Chart}
\usage{
interactive_line_chart(
  t,
  x,
  y,
  fill,
  title = "",
  est = "number",
  dec = 0,
  xtype = "Date",
  dform = "\%b-\%Y",
  breaks = NULL,
  lwidth = 1,
  color = "psrc_light"
)
}
\arguments{
\item{t}{A tibble or dataframe in long form for plotting}

\item{x}{The name of the variable you want plotted on the X-Axis}

\item{y}{The name of the variable you want plotted on the Y-Axis}

\item{fill}{The name of the variable you want the fill color of the bars to be based on}

\item{title}{Title to be used for chart, if desired - defaults to "NULL"}

\item{est}{Type for the numeric values - enter "percent", "currency" or "number", defaults to "percent"}

\item{dec}{Number of decimal points in labels - defaults to 0}

\item{xtype}{Type of values for the X-Axis either "Continuous" or "Date", defaults to "Date"}

\item{dform}{Format for Date values}

\item{breaks}{Break points to use if using a continuous scale, defaults to NULL}

\item{lwidth}{Width of lines, defaults to 1}

\item{color}{Name of color palette to use - generally defaults to "psrc_dark"}
}
\value{
line chart
}
\description{
This function allows you to create a line chart.
}
