% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrcplot_helpers.R
\name{labeller_wrap_evenly}
\alias{labeller_wrap_evenly}
\title{Helper - provide labeller with labels split into similarly sized pieces}
\usage{
labeller_wrap_evenly(width)
}
\arguments{
\item{in_string}{string to be split}

\item{wrap_width}{max string width}
}
\description{
Helper - provide labeller with labels split into similarly sized pieces
}
