% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc_charts.R
\name{line_charts}
\alias{line_charts}
\alias{static_line_chart}
\alias{interactive_line_chart}
\title{Line charts}
\usage{
static_line_chart(t, x, y, fill, ...)

interactive_line_chart(t, x, y, fill, ...)
}
\arguments{
\item{t}{A tibble in long form for plotting}

\item{x}{The name of the variable you want plotted on the X-Axis}

\item{y}{The name of the variable you want plotted on the Y-Axis}

\item{fill}{The name of the variable you want the fill color of the bars to be based on}

\item{...}{additional arguments passed to  \code{\link{generic_line}}}
}
\value{
static or interactive column or bar chart
}
\description{
Separate functions for static and interactive line charts
}
