% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrcplot_example_data.R
\docType{data}
\name{mode_share_example_data}
\alias{mode_share_example_data}
\title{Mode share example data included in psrcplot}
\format{
An object of class \code{data.frame} with 262 rows and 9 columns.
}
\usage{
mode_share_example_data
}
\description{
Mode share example data included in psrcplot
}
\keyword{datasets}
