% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc_colors.R
\name{print_psrc_palette}
\alias{print_psrc_palette}
\title{Print PSRC Color Palettes}
\usage{
print_psrc_palette()
}
\description{
Function to print to screen PSRC Color Palettes
}
\examples{

# Print the PSRC palettes to the viewer window
print_psrc_palette()

}
