% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc_colors.R
\docType{data}
\name{psrc_colors}
\alias{psrc_colors}
\title{PSRC Color Palettes}
\format{
An object of class \code{list} of length 21.
}
\usage{
psrc_colors
}
\description{
Functions to create PSRC Color Palettes
}
\details{
Available Options: "purples_inc", "oranges_inc", "greens_inc", "blues_inc", "grays_inc",
   "purples_dec", "oranges_dec", "greens_dec", "blues_dec", "grays_dec",
   "pognbgy_5", "pgnobgy_5", "gnbopgy_5", "obgnpgy_5",
   "pognbgy_10", "pgnobgy_10", "gnbopgy_10", "obgnpgy_10",
   "psrc_light", "psrc_dark", "psrc_pairs"
}
\keyword{datasets}
