% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc_colors.R
\name{psrc_print_palette}
\alias{psrc_print_palette}
\title{Print PSRC Color Palettes}
\usage{
psrc_print_palette()
}
\description{
Function to print to screen PSRC Color Palettes
}
\examples{

# Print the PSRC palettes to the viewer window
psrc_print_palette()

}
