% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc_styles.R
\name{psrc_style}
\alias{psrc_style}
\title{PSRC Styles
Useful Functions to Develop PSRC Data Visualizations}
\usage{
psrc_style()
}
\description{
This function allows you to add the PSRC theme to your ggplot graphics.
}
