% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrcplot_details.R
\docType{package}
\name{psrcplot}
\alias{psrcplot}
\alias{psrcplot-package}
\title{psrcplot: A package for easy creation of plots and maps for the Puget Sound Region.}
\description{
Various functions to create charts, tables and maps for data in the Central Puget Sound Region.
Colors, fonts and chart styles are all consistent with the design standards used by PSRC.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://psrc.github.io/psrcplot/}
}

}
\author{
Craig Helmann \email{chelmann@psrc.org}
}
