% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc_charts.R
\name{static_facet_column_chart}
\alias{static_facet_column_chart}
\title{Create Static Facet Column Charts}
\usage{
static_facet_column_chart(
  t,
  x,
  y,
  fill,
  facet,
  moe = NULL,
  est = "percent",
  ncol = 3,
  scales = "free",
  dec = 0,
  color = "psrc_dark",
  title = NULL,
  subtitle = NULL,
  source = ""
)
}
\arguments{
\item{t}{A tibble or dataframe in long form for plotting}

\item{x}{The name of the variable you want plotted on the X-Axis}

\item{y}{The name of the variable you want plotted on the Y-Axis}

\item{fill}{The name of the variable you want the fill color of the bars to be based on}

\item{facet}{The name of the variable to be the facets (facet_wrap)}

\item{moe}{The name of the variable to be used for error bars, if desired - default to "NULL"}

\item{est}{Type for the Y values - enter "percent", "currency" or "number", defaults to "percent"}

\item{ncol}{Value for the number of columns in your facet - defaults to 3}

\item{scales}{Value for axis in facets, either "fixed" or "free" - defaults to "free"}

\item{dec}{Number of decimal points in labels - defaults to 0}

\item{color}{Name of color palette to use - defaults to "psrc_dark"}

\item{title}{Title to be used for chart, if desired - defaults to "NULL"}

\item{subtitle}{Sub-title to be used for chart, if desired - defaults to "NULL"}

\item{source}{Source reference to be used for chart, if desired - defaults to blank}
}
\value{
A static facet column (vertical bar) chart; based on facet_wrap()
}
\description{
This function allows you to create facet column charts based on ggplot2's facet_wrap().
}
\examples{

library(dplyr)
# Read in the example data and filter to 2020 Population by Race by County
df <- read.csv(system.file('extdata', 'example_data.csv', package='psrcplot')) \%>\% 
      filter(Category=="Population by Race" & Year==2020) \%>\%
      filter(Geography!="Region" & Race!="Total")

my_facet <- static_facet_column_chart(t = df, 
                                      x = "Geography", 
                                      y = "share", 
                                      fill = "Geography", 
                                      facet = "Race",
                                      moe = 'share_moe',
                                      ncol = 4,
                                      scales = "fixed",
                                      color = "pgnobgy_5",
                                      title = "Something",
                                      subtitle = "Something something",
                                      source = paste("Source: ACS 5-Year Estimates, table B3002",
                                                 "for King, Kitsap, Pierce and Snohomish counties.",
                                                 sep = "\n"))

}
