% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrcplot_example_data.R
\docType{data}
\name{summary_pums_example_data}
\alias{summary_pums_example_data}
\title{Summary PUMS example data included in psrcplot}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 10 rows and 5 columns.
}
\usage{
summary_pums_example_data
}
\description{
see frequently-asked-questions.Rmd vignette for script
}
\keyword{datasets}
