% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrcplot_example_data.R
\docType{data}
\name{transit_access_example_data}
\alias{transit_access_example_data}
\title{Transit access example data included in psrcplot}
\description{
Transit access example data included in psrcplot
}
