% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrcplot_helpers.R
\name{add_citation}
\alias{add_citation}
\title{Helper - add source citation to plotly object}
\usage{
add_citation(p, source)
}
\arguments{
\item{p}{plotly object}

\item{source}{Source reference as character string}
}
\description{
Helper - add source citation to plotly object
}
