% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc_charts.R
\name{create_bar_chart}
\alias{create_bar_chart}
\title{Create PSRC Bar Chart}
\usage{
create_bar_chart(
  t,
  w.x,
  w.y,
  f,
  h.ref = NULL,
  h.ref.nm = NULL,
  h.ref.cl = NULL,
  w.moe = NULL,
  w.title = NULL,
  w.sub.title = NULL,
  w.pos = "dodge",
  est.type = "percent",
  w.dec = 0,
  w.color = "psrc_dark",
  w.interactive = "no"
)
}
\arguments{
\item{t}{A tibble or dataframe in long form for plotting}

\item{w.x}{The name of the variable you want plotted on the X-Axis}

\item{w.y}{The name of the variable you want plotted on the Y-Axis}

\item{f}{The name of the variable you want the fill color of the bars to be based on}

\item{h.ref}{A list of values to be used for any horizontal reference lines - default is "NULL"}

\item{h.ref.nm}{A list of names to be used for any horizontal reference lines - default is "NULL"}

\item{h.ref.cl}{A list of colors to be used for any horizontal reference lines - default is "NULL"}

\item{w.moe}{The name of the variable to be used for error bars, if desired - default is "NULL"}

\item{w.title}{Title to be used for chart, if desired - defaults to "NULL"}

\item{w.sub.title}{Sub-title to be used for chart, if desired - defaults to "NULL"}

\item{w.pos}{Determines if the bars are side-by-side(dodge) or stacked(stack) - defaults to "dodge"}

\item{est.type}{Type for the Y values - enter "percent", "currency" or "number", defaults to "percent"}

\item{w.dec}{Number of decimal points in labels - defaults to 0}

\item{w.color}{Name of color palette to use - defaults to "psrc_dark"}

\item{w.interactive}{Enable hover text and other interactive features - defaults to "no"}
}
\value{
bar chart that is either static or interactive depending on choice
}
\description{
This function allows you to create bar charts.
}
\examples{

library(dplyr)
library(psrccensus)

# Modes
total.work.trips <- c("B08006_001")
da.work.trips <- c("B08006_003")
cp.work.trips <- c("B08006_004")
trn.work.trips <- c("B08006_008")
bke.work.trips <- c("B08006_014")
wlk.work.trips <- c("B08006_015")
oth.work.trips <- c("B08006_016")
wfh.work.trips <- c("B08006_017")

all.trips <- c(total.work.trips, da.work.trips, 
               cp.work.trips, trn.work.trips, 
               bke.work.trips, wlk.work.trips, 
               oth.work.trips, wfh.work.trips)
               
non.da.trips <- c(cp.work.trips, trn.work.trips,
                  bke.work.trips, wlk.work.trips,
                  wfh.work.trips)

rtp <- 0.49
fed <- 0.333

work.mode <- get_acs_recs(geography = 'county',
                          table.names = c('B08006'), 
                          years=c(2010,2015,2020), 
                          acs.type = 'acs5') \%>\% 
             filter(variable \%in\% all.trips)
             
mode.totals <- work.mode \%>\% 
               filter(variable \%in\% total.work.trips) \%>\% 
               select(name, year, estimate) \%>\% 
               rename(total=estimate)
               
work.mode <- left_join(work.mode, mode.totals, by=c("name","year")) \%>\% 
             mutate(share=estimate/total)
             
tbl <- work.mode \%>\%
       filter(variable \%in\% non.da.trips) \%>\%
       filter(name=="Region") \%>\%
       select(name, year, share) \%>\%
       group_by(name, year) \%>\%
       summarise(share=sum(share)) \%>\%
       as_tibble() \%>\%
       mutate(year=as.character(year))


# Chart with Reference Lines        
ms.chart <- create_bar_chart(t=tbl, w.x="year", w.y="share", f="name",
                             h.ref = c(rtp,fed), 
                             h.ref.nm = c("2050 RTP Forecast","Federal Performance Target"),
                             h.ref.cl = psrc_colors$PrGn,
                             w.title= "Non-SOV Mode Share to Work", 
                             w.sub.title="ACS 5yr Data Table B08006")

}
