% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc_charts.R
\name{create_bar_chart}
\alias{create_bar_chart}
\title{Create PSRC Bar Chart}
\usage{
create_bar_chart(
  t,
  w.x,
  w.y,
  f,
  w.moe = NULL,
  w.title = NULL,
  w.sub.title = NULL,
  w.pos = "dodge",
  est.type = "percent",
  w.dec = 0,
  w.color = "psrc_dark",
  w.interactive = "no"
)
}
\arguments{
\item{t}{A tibble or dataframe in long form for plotting}

\item{w.x}{The name of the variable you want plotted on the X-Axis}

\item{w.y}{The name of the variable you want plotted on the Y-Axis}

\item{f}{The name of the variable you want the fill color of the bars to be based on}

\item{w.moe}{The name of the variable to be used for error bars, if desired - default to "NULL"}

\item{w.title}{Title to be used for chart, if desired - defaults to "NULL"}

\item{w.sub.title}{Sub-title to be used for chart, if desired - defaults to "NULL"}

\item{w.pos}{Determines if the bars are side-by-side(dodge) or stacked(stack) - defaults to "dodge"}

\item{est.type}{Type for the Y values - enter "percent", "currency" or "number", defaults to "percent"}

\item{w.dec}{Number of decimal points in labels - defaults to 0}

\item{w.color}{Name of color palette to use - defaults to "psrc_dark"}

\item{w.interactive}{Enable hover text and other interactive features - defaults to "no"}
}
\value{
bar chart that is either static or interactive depending on choice
}
\description{
This function allows you to create bar charts.
}
\examples{
\dontrun{
library(dplyr)

# Read in the example data and filter to Mode to Work for Everyone for all years in the data
df <- read.csv("inst/extdata/example_data.csv") \%>\% 
      filter(Category=="Mode to Work by Race") \%>\%
      filter(Geography=="Region" & Race=="Total") \%>\%
      mutate(Year = as.character(Year))

# Create a chart for mode shares by year with error bars
my.chart <- create_bar_chart(t=df, w.x="Mode", w.y="share", f="Year", w.moe="share_moe", 
                             w.color="psrc_light")

# Create a chart for mode shares by year without error bars
my.chart <- create_bar_chart(t=df, w.x="Mode", w.y="share", f="Year", 
                             w.color="psrc_light")
                             
# Create a chart for mode shares by year without error bars with Titles
my.chart <- create_bar_chart(t=df, w.x="Mode", w.y="share", f="Year", 
                             w.color="psrc_light",
                             w.title="Mode Share to Work",
                             w.sub.title="by Census Year")
                             
# Create a chart for mode shares by year without error bars with Titles and make interactive
my.chart <- create_bar_chart(t=df, w.x="Mode", w.y="share", f="Year", 
                             w.color="psrc_light",
                             w.title="Mode Share to Work",
                             w.sub.title="by Census Year",
                             w.interactive="yes")   
                                                        
# Create a chart for mode shares by year with error bars with Main Title only and make interactive
my.chart <- create_bar_chart(t=df, w.x="Mode", w.y="share", f="Year", 
                             w.color="psrc_light",
                             w.title="Mode Share to Work",
                             w.moe="share_moe",
                             w.interactive="yes") 

}

}
