% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc_bubble_charts.R
\name{create_bubble_chart}
\alias{create_bubble_chart}
\title{Create PSRC Bubble Chart}
\usage{
create_bubble_chart(
  t,
  x,
  y,
  fill,
  s,
  color = "psrc_light",
  title = NULL,
  subtitle = NULL
)
}
\arguments{
\item{t}{A tibble or dataframe in long form for plotting}

\item{x}{The name of the variable you want plotted on the X-Axis}

\item{y}{The name of the variable you want plotted on the Y-Axis}

\item{fill}{The name of the variable you want the fill color of the bubbles to be based on}

\item{s}{The name of the variable used to size the bubbles}

\item{color}{Name of color palette to use - defaults to "psrc_light"}

\item{title}{Title to be used for chart, if desired - defaults to "NULL"}

\item{subtitle}{Sub-title to be used for chart, if desired - defaults to "NULL"}
}
\value{
bubble chart
}
\description{
This function allows you to create a bubble charts.
}
\author{
Craig Helmann
}
