% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc_charts.R
\name{create_facet_bar_chart}
\alias{create_facet_bar_chart}
\title{Create PSRC Charts}
\usage{
create_facet_bar_chart(
  t,
  w.x,
  w.y,
  f,
  g,
  est.type = "percent",
  w.scales = "free",
  w.facet = 3,
  w.dec = 0,
  l.pos = "above",
  w.color = "psrc_distinct_10"
)
}
\arguments{
\item{t}{A tibble or dataframe in long form for plotting}

\item{w.x}{The name of the variable you want plotted on the X-Axis}

\item{w.y}{The name of the variable you want plotted on the Y-Axis}

\item{f}{The name of the variable you want the fill color of the bars to be based on}

\item{g}{The name of the variable to be the facets}

\item{est.type}{Type for the Y values - enter "percent", "currency" or "number", default in "percent"}

\item{w.scales}{Value for axis in facets, either "fixed" or "free" - defaults to "free"}

\item{w.facet}{Value for the number of columns in your facet - defaults to 3}

\item{w.dec}{Number of decimal points in labels - defaults to 0}

\item{l.pos}{Position for the bar labels of "above" or "within" - defaults to "above"}

\item{w.color}{Name of color palette to use - defaults to "psrc_distinct_10"}
}
\value{
facet bar chart that is either static or interactive depending on choice
}
\description{
This function allows you to create facet bar charts.
}
