% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc_charts.R
\name{create_facet_bar_chart}
\alias{create_facet_bar_chart}
\title{Create PSRC Facet Bar Charts}
\usage{
create_facet_bar_chart(
  t,
  x,
  y,
  fill,
  g,
  moe = NULL,
  est = "percent",
  scales = "free",
  facet = 3,
  dec = 0,
  lpos = "above",
  color = "psrc_dark",
  title = NULL,
  subtitle = NULL,
  interactive = "no"
)
}
\arguments{
\item{t}{A tibble or dataframe in long form for plotting}

\item{x}{The name of the variable you want plotted on the X-Axis}

\item{y}{The name of the variable you want plotted on the Y-Axis}

\item{fill}{The name of the variable you want the fill color of the bars to be based on}

\item{g}{The name of the variable to be the facets}

\item{moe}{The name of the variable to be used for error bars, if desired - default to "NULL"}

\item{est}{Type for the Y values - enter "percent", "currency" or "number", defaults to "percent"}

\item{scales}{Value for axis in facets, either "fixed" or "free" - defaults to "free"}

\item{facet}{Value for the number of columns in your facet - defaults to 3}

\item{dec}{Number of decimal points in labels - defaults to 0}

\item{lpos}{Position for the bar labels of "above" or "within" - defaults to "above"}

\item{color}{Name of color palette to use - defaults to "psrc_dark"}

\item{title}{Title to be used for chart, if desired - defaults to "NULL"}

\item{subtitle}{Sub-title to be used for chart, if desired - defaults to "NULL"}

\item{interactive}{Enable hover text and other interactive features - defaults to "no"}
}
\value{
facet bar chart that is either static or interactive depending on choice
}
\description{
This function allows you to create facet bar charts.
}
\examples{

library(dplyr)
# Read in the example data and filter to 2020 Population by Race by County
df <- psrcplot::mode_share_example_data \%>\% 
      filter(Category=="Population by Race" & Year==2020) \%>\%
      filter(Geography!="Region" & Race!="Total")

# Create a facet chart for population by race using counties as the facet
my.chart <- create_facet_bar_chart(t=df, x="Race", y="share", 
                                   fill="Race", g="Geography", 
                                   facet=2, scales="fixed")

}
