% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc_charts.R
\name{create_treemap_chart}
\alias{create_treemap_chart}
\title{Create PSRC TreeMap Chart}
\usage{
create_treemap_chart(
  t,
  w.area,
  w.fill,
  w.title = NULL,
  w.sub.title = NULL,
  est.type = "percent",
  w.dec = 0,
  w.color = "psrc_light"
)
}
\arguments{
\item{t}{A tibble or dataframe in long form for plotting}

\item{w.area}{The name of the variable with thee value you want to use to size the bars}

\item{w.fill}{The name of the variable you want to fill the bars}

\item{w.title}{Title to be used for chart, if desired - defaults to "NULL"}

\item{w.sub.title}{Sub-title to be used for chart, if desired - defaults to "NULL"}

\item{est.type}{Type for the Y values - enter "percent", "currency" or "number", defaults to "percent"}

\item{w.dec}{Number of decimal points in labels - defaults to 0}

\item{w.color}{Name of color palette to use - defaults to "psrc_light"}
}
\value{
static treemap chart
}
\description{
This function allows you to create bar charts.
}
\examples{
\dontrun{
library(dplyr)

# Read in the example data and filter to Mode to Work for Everyone for all years in the data
df <- read.csv("inst/extdata/example_data.csv") \%>\% 
      filter(Category=="Mode to Work by Race") \%>\%
      filter(Geography=="Region" & Race=="Total") \%>\%
      mutate(Year = as.character(Year)) \%>\%
      filter(Year=="2020")

my.chart <- create_treemap_chart(t=df, w.area="share", w.fill="Mode", w.title="Mode Share to Work")

}

}
