% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc_echarts.R
\name{echart_bar_chart}
\alias{echart_bar_chart}
\title{echart4r Bar Chart}
\usage{
echart_bar_chart(
  t,
  x,
  y,
  est,
  fill,
  pos = "NULL",
  column_vs_bar = "column",
  ...
)
}
\arguments{
\item{t}{A data frame in long form for plotting}

\item{x}{The name of the category variable}

\item{y}{The name of the variable with numeric values to plot}

\item{est}{Select "number", "percent", or "currency" - defaults to "percent"}

\item{fill}{The name of the variable you want the fill color of the bars or lines to be based on}

\item{pos}{Determines if the bars are side-by-side(NULL) or stacked("grp") - defaults to NULL}

\item{column_vs_bar}{Select "column" or "bar" for chart orientation - defaults to "column"}

\item{...}{additional arguments passed to  \code{\link{generic_echart}}}
}
\value{
An interactive bar or column chart via echarts4r
}
\description{
echart4r Bar Chart
}
\examples{
library(tidyverse)

d <- mode_share_example_data |>
 filter(Year == 2020 & Category == 'Population by Race') |>
 mutate(Race = str_wrap(Race, 20))

color_palette <- c('#8dd3c7','#ffffb3','#bebada','#fb8072','#80b1d3')

echart_bar_chart(t = d,
                x = "Race",
                y = "share",
                fill = "Geography",
                color = color_palette,
                title = "Population by Race",
                subtitle = "In 2020",
                pos = NULL,
                column_vs_bar = "bar",
                est = "number")

}
\author{
Christy Lam
}
