% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc_echarts.R
\name{echart_line_chart}
\alias{echart_line_chart}
\title{echart4r Line Chart}
\usage{
echart_line_chart(t, x, y, est, fill, ...)
}
\arguments{
\item{t}{A data frame in long form for plotting}

\item{x}{The name of the category variable}

\item{y}{The name of the variable with numeric values to plot}

\item{est}{Select "number", "percent", or "currency" - defaults to "percent"}

\item{fill}{The name of the variable you want the fill color of the bars or lines to be based on}

\item{...}{Additional arguments passed to \code{\link{generic_echart}}}
}
\value{
An interactive line chart via echarts4r
}
\description{
echart4r Line Chart
}
