% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc_line_charts.R
\name{generic_line}
\alias{generic_line}
\title{Generic Line Chart}
\usage{
generic_line(
  t,
  x,
  y,
  fill,
  est = NULL,
  text = NULL,
  dec = 0,
  dform = "\%b-\%Y",
  breaks = NULL,
  lwidth = 1,
  color = "gnbopgy_5",
  title = NULL,
  subtitle = NULL,
  source = "",
  alt = NULL,
  xlabel = NULL,
  ylabel = NULL,
  interactive = FALSE
)
}
\arguments{
\item{t}{A tibble or dataframe in long form for plotting}

\item{x}{The name of the variable you want plotted on the X-Axis}

\item{y}{The name of the variable you want plotted on the Y-Axis}

\item{fill}{The name of the variable you want the fill color of the bars to be based on}

\item{est}{Type for the numeric values - enter "percent", "currency" or "number", defaults to "percent"}

\item{text}{name of the variable to use as tooltip}

\item{dec}{Number of decimal points in labels - defaults to 0}

\item{dform}{Format for Date values}

\item{breaks}{Break points for a continuous scale or date break description for date scale; default NULL}

\item{lwidth}{Width of lines, defaults to 1}

\item{color}{Name of color palette to use - generally defaults to "pgnobgy_5"}

\item{title}{Title to be used for chart, if desired - defaults to "NULL"}

\item{subtitle}{Sub-title to be used for chart, if desired - defaults to "NULL"}

\item{source}{Source reference as character string, if desired - defaults to blank}

\item{alt}{Text to be used for alt-text, if desired - defaults to "NULL"}

\item{xlabel}{category-axis title to be used for chart, if desired - defaults to "NULL"}

\item{ylabel}{numeric-axis title to be used for chart, if desired - defaults to "NULL"}

\item{interactive}{Enable hover text and other interactive features - defaults to FALSE}
}
\value{
line chart
}
\description{
Used for both static and interactive line functions.
}
\author{
Craig Helmann, Michael Jensen
}
