% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrcplot_helpers.R
\name{labeller_wrap_evenly}
\alias{labeller_wrap_evenly}
\title{wrap_evenly function in form to apply to ggplot2 labeller spec}
\usage{
labeller_wrap_evenly(width)
}
\arguments{
\item{width}{max string width}
}
\description{
wrap_evenly function in form to apply to ggplot2 labeller spec
}
