% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrcplot_helpers.R
\name{make_interactive}
\alias{make_interactive}
\title{Helper - make static ggplot object an interactive plotly object}
\usage{
make_interactive(p, title = NULL, subtitle = NULL)
}
\arguments{
\item{p}{ggplot object}

\item{title}{Title to be used for chart, if desired - defaults to "NULL"}

\item{subtitle}{Sub-title to be used for chart, if desired - defaults to "NULL"}
}
\value{
plotly object
}
\description{
Helper - make static ggplot object an interactive plotly object
}
\author{
Craig Helmann, Michael Jensen
}
