% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc_colors.R
\docType{data}
\name{psrc_colors}
\alias{psrc_colors}
\title{PSRC Color Palettes}
\format{
An object of class \code{list} of length 31.
}
\usage{
psrc_colors
}
\description{
Functions to create PSRC Color Palettes
}
\keyword{datasets}
