% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrcplot_details.R
\docType{package}
\name{psrcplot}
\alias{psrcplot}
\alias{psrcplot-package}
\title{psrcplot: A package for easy creation of plots and maps for the Puget Sound Region.}
\description{
Various functions to create charts, tables and maps for data in the Central Puget Sund Region.
This package is currently under development and will be updated frequently over the last half of 2022.
Colors, fonts and chart styles are all consistent with teh design standards used by PSRC.
}
\author{
Suzanne Childress \email{schildress@psrc.org}
}
