% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc_colors.R
\name{scale_fill_continuous_psrc}
\alias{scale_fill_continuous_psrc}
\title{PSRC Continuous Color Palettes}
\usage{
scale_fill_continuous_psrc(name)
}
\arguments{
\item{name}{The name of the continuous color palette}
}
\description{
Function for a continuous color fill for use in ggplot2
}
