% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc_colors.R
\name{scale_fill_discrete_psrc}
\alias{scale_fill_discrete_psrc}
\title{PSRC Discrete Color Palettes}
\usage{
scale_fill_discrete_psrc(name)
}
\description{
Function for a discrete color fill for use in ggplot2
}
