% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrcplot_helpers.R
\name{shared_params}
\alias{shared_params}
\title{Parameters shared among other functions}
\arguments{
\item{fill}{The name of the variable you want the fill color of the bars to be based on}

\item{est}{Type for the numeric values - enter "percent", "currency" or "number", defaults to "percent"}

\item{dec}{Number of decimal points in labels - defaults to 0}

\item{color}{Name of color palette to use - generally defaults to "pgnobgy_5"}

\item{title}{Title to be used for chart, if desired - defaults to "NULL"}

\item{subtitle}{Sub-title to be used for chart, if desired - defaults to "NULL"}

\item{source}{Source reference as character string, if desired - defaults to blank}
}
\description{
Parameters shared among other functions
}
