% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc_line_charts.R
\name{static_facet_line_chart}
\alias{static_facet_line_chart}
\title{Facet Line Chart}
\usage{
static_facet_line_chart(t, x, y, fill, facet, scales = "free", ncol = 3, ...)
}
\arguments{
\item{t}{A tibble or dataframe in long form for plotting}

\item{x}{The name of the variable you want plotted on the X-Axis}

\item{y}{The name of the variable you want plotted on the Y-Axis}

\item{fill}{The name of the variable you want the fill color of the bars to be based on}

\item{facet}{The name of the variable to be the facets}

\item{scales}{Value for axis in facets, either "fixed" or "free" - defaults to "free"}

\item{ncol}{Value for the number of columns in your facet - defaults to 3}

\item{...}{additional arguments passed to  \code{\link{generic_line}}}
}
\value{
A static facet line chart; based on facet_wrap()
}
\description{
Facet Line Chart
}
