% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrcplot_helpers.R
\name{wrap_evenly}
\alias{wrap_evenly}
\title{Make similiarly sized multi-line strings from single-line strings}
\usage{
wrap_evenly(str_vector, wrap_width)
}
\arguments{
\item{str_vector}{string or character vector to insert line breaks}

\item{wrap_width}{max string width}
}
\value{
identical string or character vector, evenly wrapped
}
\description{
Make similiarly sized multi-line strings from single-line strings
}
\author{
Michael Jensen
}
