% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrcplot_helpers.R
\name{wrap_labels_evenly}
\alias{wrap_labels_evenly}
\title{Helper - split labels into similarly sized pieces}
\usage{
wrap_labels_evenly(width)
}
\arguments{
\item{width}{max string width}
}
\description{
Helper - split labels into similarly sized pieces
}
